8237192: Generate stripped/public pdbs on Windows for jdk images Reviewed-by: erikj, ihse Contributed-by: christoph.langer@sap.com, matthias.baesken@sap.com
(cherry picked from commit a4177910a972f75cdf89541fb290765ae648f923)
